// frontend/src/api/generated/models/plannedWorkoutDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */
import type { SportType } from './sportType';
import type { WorkoutSegmentDto } from './workoutSegmentDto';

export interface PlannedWorkoutDto {
  id?: string;
  trainingWeekId?: string;
  date?: string;
  sport?: SportType;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  coachDescription?: string | null;
  isRestDay?: boolean;
  /** @nullable */
  plannedDistanceMeters?: number | null;
  /** @nullable */
  plannedDurationSeconds?: number | null;
  /** @nullable */
  plannedElevationGainMeters?: number | null;
  /** @nullable */
  segments?: WorkoutSegmentDto[] | null;
}